+++ fj-doc-maven-plugin/src/test/java/test/org/fugerit/java/doc/project/facade/TestVenusContext.java
@@ -10,12 +10,20 @@ import java.io.File;
 @Slf4j
 public class TestVenusContext {
 
+    private static final File TARGET = new File( "target/" );
+
     @Test
     public void testVenusContext() {
-        File folder = new File( "target/" );
-        VenusContext context = new VenusContext( folder, VenusContext.VERSION_NA_FULL_PROCESS, "fj-doc-base" );
+        VenusContext context = new VenusContext( TARGET, VenusContext.VERSION_NA_FULL_PROCESS, "fj-doc-base" );
         Assert.assertNotNull( context.getTestResourcesFolder() );
         Assert.assertNotNull( context.getMainResourcesFolder() );
+        Assert.assertFalse( context.isAsciidocFreemarkerHandlerAvailable() );
+    }
+
+    @Test
+    public void testAsciiDocCheck() {
+        VenusContext context = new VenusContext( TARGET, "8.10.8", "fj-doc-base" );
+        Assert.assertTrue( context.isAsciidocFreemarkerHandlerAvailable() );
     }
 
 }